+++ integration/executor/src/main/resources/tbmq-integration-executor.yml
@@ -38,14 +38,15 @@ queue:
       retries: "${TB_IE_MSG_ACK_STRATEGY_RETRIES:5}"
       # Time to wait in the consumer thread before retrying (in s).
       pause-between-retries: "${TB_IE_MSG_ACK_STRATEGY_PAUSE_BETWEEN_RETRIES:1}"
+  integration-event:
     # Poll interval for the dedicated 'tbmq.ie.event' lifecycle-event topics (in ms).
-    event-poll-interval: "${TB_IE_EVENT_MSG_POLL_INTERVAL:1000}"
+    poll-interval: "${TB_IE_EVENT_MSG_POLL_INTERVAL:1000}"
     # Timeout for processing a pack of lifecycle events (in ms).
-    event-pack-processing-timeout: "${TB_IE_EVENT_MSG_PACK_PROCESSING_TIMEOUT:30000}"
-    event-ack-strategy:
+    pack-processing-timeout: "${TB_IE_EVENT_MSG_PACK_PROCESSING_TIMEOUT:30000}"
+    ack-strategy:
       # Processing strategy for 'tbmq.ie.event' topics. Accepted values: SKIP_ALL, RETRY_ALL.
       # Defaults to SKIP_ALL because lifecycle events are best-effort hints - retrying would redeliver
-      # stale events out of order; kept independent from the data stream so it is not coupled to RETRY_ALL.
+      # stale events out of order; kept independent of the data stream, so it is not coupled to RETRY_ALL.
       type: "${TB_IE_EVENT_MSG_ACK_STRATEGY_TYPE:SKIP_ALL}"
       # Number of retries. Use 0 for unlimited. Applies to the RETRY_ALL strategy.
       retries: "${TB_IE_EVENT_MSG_ACK_STRATEGY_RETRIES:5}"
@@ -171,14 +172,17 @@ queue:
       additional-consumer-config: "${TB_KAFKA_IE_MSG_ADDITIONAL_CONSUMER_CONFIG:max.poll.records:50}"
       # Additional Kafka producer configs separated by semicolon for the `tbmq.msg.ie` topics.
       additional-producer-config: "${TB_KAFKA_IE_MSG_ADDITIONAL_PRODUCER_CONFIG:}"
+    integration-event:
       # Kafka topic properties separated by semicolon for the dedicated `tbmq.ie.event` lifecycle-event topics.
       # Retention defaults shorter than the data topic since events are best-effort hints.
-      event-topic-properties: "${TB_KAFKA_IE_EVENT_MSG_TOPIC_PROPERTIES:retention.ms:86400000;segment.bytes:26214400;retention.bytes:104857600;partitions:1;replication.factor:1}"
+      topic-properties: "${TB_KAFKA_IE_EVENT_MSG_TOPIC_PROPERTIES:retention.ms:86400000;segment.bytes:26214400;retention.bytes:104857600;partitions:1;replication.factor:1}"
       # Additional Kafka consumer configs separated by semicolon for the `tbmq.ie.event` topics.
-      additional-event-consumer-config: "${TB_KAFKA_IE_EVENT_MSG_ADDITIONAL_CONSUMER_CONFIG:max.poll.records:50}"
-      # Additional Kafka producer configs separated by semicolon for the `tbmq.ie.event` topics. Configured non-blocking
-      # (low max.block.ms); used only by the broker, present here for shared-settings parity.
-      additional-event-producer-config: "${TB_KAFKA_IE_EVENT_MSG_ADDITIONAL_PRODUCER_CONFIG:max.block.ms:5000}"
+      additional-consumer-config: "${TB_KAFKA_IE_EVENT_MSG_ADDITIONAL_CONSUMER_CONFIG:max.poll.records:50}"
+      # Additional Kafka producer configs separated by semicolon for the `tbmq.ie.event` topics. Lifecycle events are
+      # sent synchronously on the MQTT processing thread (same path as a regular publish), so a slow or unavailable Kafka
+      # can stall that thread for up to max.block.ms before the send fails, after which the event is dropped and counted.
+      # Lower max.block.ms to tighten that bound at the cost of more dropped events under Kafka pressure.
+      additional-producer-config: "${TB_KAFKA_IE_EVENT_MSG_ADDITIONAL_PRODUCER_CONFIG:max.block.ms:5000}"
     # Common prefix for all Kafka topics, producers, consumer groups, and consumers. Defaults to empty string, meaning no prefix is applied.
     kafka-prefix: "${TB_KAFKA_PREFIX:}"
     # Custom consumer configuration per Kafka topic.
